Move tests to the right place
